# budgetflow-backend/swagger.yaml
openapi: 3.0.0
info:
  title: BudgetFlow API 명세서 (MVP)
  version: 1.0.0
  description: BudgetFlow 단일 백엔드 서버의 API 명세서입니다.
servers:
  - url: http://localhost:3000/api
    description: 로컬 개발 서버
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Project:
      type: object
      properties:
        id:
          type: string
          example: "proj_01"
        organizationId:
          type: string
          example: "org_inha_cs_2026"
        name:
          type: string
          example: "2026 상반기 팀 프로젝트"
        totalBudget:
          type: number
          example: 500000
        status:
          type: string
          enum: [active, closed]
          example: "active"
        slackChannelId:
          type: string
          example: "C08ABC123"
        slackChannelName:
          type: string
          example: "budgetflow-demo"
        templateFileName:
          type: string
          nullable: true
          example: "template_2026.xlsx"
        templateMappingStatus:
          type: string
          enum: [none, suggested, confirmed]
          example: "confirmed"
        createdAt:
          type: string
          example: "2026-05-01T09:00:00Z"
        closedAt:
          type: string
          nullable: true
          example: null
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "UNAUTHORIZED"
            message:
              type: string
              example: "인증 토큰이 누락되었습니다."
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
security:
  - bearerAuth: []
paths:
...
  /projects:
    get:
      summary: 프로젝트 목록 조회
      tags:
        - Project
      responses:
        '200':
          description: active 우선, createdAt 최신순
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
              example:
                - id: "proj_01"
                  organizationId: "org_inha_cs_2026"
                  name: "2026 상반기 팀 프로젝트"
                  totalBudget: 500000
                  status: "active"
                  slackChannelId: "C08ABC123"
                  slackChannelName: "budgetflow-demo"
                  templateFileName: "template_2026.xlsx"
                  templateMappingStatus: "confirmed"
                  createdAt: "2026-05-01T09:00:00Z"
                  closedAt: null
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: 프로젝트 생성
      tags:
        - Project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [organizationId, name, totalBudget, slackChannelName]
              properties:
                organizationId:
                  type: string
                  example: "org_inha_cs_2026"
                name:
                  type: string
                  example: "2026 하반기 행사"
                totalBudget:
                  type: number
                  example: 300000
                slackChannelName:
                  type: string
                  example: "budgetflow-2nd"
                templateFileName:
                  type: string
                  example: "template_v2.xlsx"
      responses:
        '201':
          description: 생성된 프로젝트 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
  /projects/{projectId}:
    get:
      summary: 프로젝트 단건 조회
      tags:
        - Project
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: 프로젝트 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{projectId}/close:
    post:
      summary: 정산 마감
      tags:
        - Project
      parameters:
        - $ref: '#/components/parameters/projectId'
      responses:
        '200':
          description: 마감된 프로젝트 반환 (status=closed)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              example:
                id: "proj_01"
                status: "closed"
                closedAt: "2026-05-29T10:00:00Z"

  # ──────────────────────────────────────────
  # Expense API
  # ──────────────────────────────────────────
...
components:
  parameters:
    projectId:
      in: path
      name: projectId
      required: true
      schema:
        type: string
      example: "proj_01"
...
  responses:
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "인증 토큰이 누락되었습니다."
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "NOT_FOUND"
              message: "요청한 리소스를 찾을 수 없습니다."
    BadRequest:
      description: 입력 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: "VALIDATION_ERROR"
              message: "입력값이 올바르지 않습니다."
              details:
                - field: "name"
                  message: "2자 이상 입력해야 합니다."